+++ pkgtmpl/python/_pkgname_/metadata.py
@@ -1,3 +1,2 @@
-"""Description for {{ appname }} goes here"""
 __version_info__ = (0, 1, 0)
 __version__ = '.'.join(map(str, __version_info__))